requirements.txt
# Core ML dependencies
diffusers>=0.21.0
transformers>=4.21.0
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.12.0
safetensors>=0.3.0

# Image processing
Pillow>=9.0.0

# API framework
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
python-multipart
aiofiles

# Utilities
requests
numpy
scipy

# Development and testing
pytest>=7.0.0
pytest-asyncio
httpx
python-dotenv

# Optional performance improvements
xformers 